Move DirtySanta fixup to full init phase, remove from primary runtime image
Searching found the smem initialization is done during the
core_initcall() phase.  With that information the earliest the DirtySanta
smem fixup can be done is subsys_initcall() time.

The last ingredient is figuring out the prototype for initcall functions.
With this done the functions can be __init and the data __initdata.

The hooks into crucial IPA functions have been removed.  Everything is
done in the init timeframe.  This is a much cleaner implementation for
this workaround.