DirtySanta fixup v3, available via sysfs
This version introduces a new interface for changing the crucial SMEM
values.  The values can not be accessed via sysfs.  If the values aren't
set via kernel command-line or kernel configuration, the values in sysfs
will be writeable.  Beware the values MUST be set before the modem
subsystem is started (pm-service); otherwise the modem firmware WILL
panic, causing SSRs and often leading to kernel panic.

On the intended target the key files appear as
"/sys/devices/soc/2080000.qcom,mss/dirtysanta_lg_model_name"
and "/sys/devices/soc/2080000.qcom,mss/dirtysanta_sim_num".

Done for LineageOS as this type of interface is acceptable to LineageOS,
whereas patching the kernel command-line in the boot image is not.